+++ CHANGES.rst
@@ -9,7 +9,8 @@ Unreleased
 - Fix `model_to_dict` to detect changes in a parent model when using
   `inherit=True` (backwards-incompatible for users who were directly
   using previous version) (gh-576)
-- Use an iterator for `clean_duplicate_history`
+- Use an iterator for `clean_duplicate_history` (gh-604)
+- Add support for Python 3.8 and Django 3.0 (gh-610)
 
 2.7.3 (2019-07-15)
 ------------------